feat: Update VideosViewSet to filter by category
This commit updates the VideosViewSet in the frmlbsite/views.py file to include a new method, get_queryset, which filters the videos based on the provided category query parameter. If a category is specified in the request, only videos belonging to that category will be returned. If no category is provided, all videos will be returned. This change improves the functionality of the API by allowing users to retrieve videos based on specific categories.